added new welcome card

+++ src/store/modules/glyph/mutations.js
@@ -1,7 +1,7 @@
 // import Vue from 'vue'
 import paper from 'paper'
 import Vue from 'vue'
-import DrawingBox from '../../../glyphs/DrawingBox';
+import DrawingBox from '../../../glyphs/DrawingBox'
 
 function glyphScope() {
   for (let i = 0; i < 3; i++) {
@@ -29,7 +29,7 @@ const checkRect = boundingRect => {
 export default {
   setLayersUp: (state, maxDisplayedGlyphs) => {
     for (let i = 0; i < maxDisplayedGlyphs; i++) {
-      let newPaperLayer = new paper.Layer({name: `glyph${i}`}) // automatically attached to project
+      const newPaperLayer = new paper.Layer({name: `glyph${i}`}) // automatically attached to project
       newPaperLayer.opacity = 1
     }
     state.activeLayer = 0
@@ -102,16 +102,22 @@ export default {
     state.project.bindings = bindings
   },
 
+  deleteBinding: (state, binding) => {
+    state.project.bindings = state.project.bindings.filter(
+        binding_ => !(binding_.element === binding.element && binding_.name === binding.name)
+    )
+  },
+
   updateGlyphNames: (state) => {
     state.totalGlyphNum = state.project.glyphs.reduce((total, newGlyph) => total + [...newGlyph.iter()].length, 0)
     state.project.glyphNames = [...state.project.glyphs.reduce((names, newGlyph) => {
-      for (let glyph of [...newGlyph.iter(true)]) {
+      for (const glyph of [...newGlyph.iter(true)]) {
         names.add(glyph.name)
       }
       return names
     }, new Set())]
     state.project.mainGlyphNames = [...state.project.glyphs.reduce((names, newGlyph) => {
-      for (let glyph of [...newGlyph.iter(true)]) {
+      for (const glyph of [...newGlyph.iter(true)]) {
         if (glyph.shape === glyph.constructor.shapes.main) {
           names.add(glyph.name)
         }
@@ -129,29 +135,17 @@ export default {
     state.glyphElements = glyphClass.elements // update glyph elements for feature binding
     // extract naming field from data-points and initialise glyphs in order of appearance in file
     // eslint-disable-next-line new-cap
-    state.project.glyphs = parsedData.map((dataPoint, layerIndex) => new glyphClass(
-        layerIndex,
-        namingField === '_default' ? layerIndex : dataPoint[namingField],  // use layer number as id if no naming field was specified
-        glyphName)
-    )
-  },
-
-  makeEmptyGlyphs: (state, {glyphName, createOptions}) => {
-    if (state.project.glyphs.length === 0) {
-      throw Error("Databound glyphs must be added before empty glyphs can be created")
-    }
-    const glyphClass = state.glyphTypes.find(glyphType => glyphType.type.startsWith(state.glyphTypeName))
-    if (typeof glyphClass === 'undefined') {
-      throw Error(`Unknown glyph type: ${state.glyphTypeName}`)
-    }
-    for (let glyph of state.project.glyphs) {
-      let emptyGlyph
-      if (createOptions) {
-        emptyGlyph = new glyphClass(glyph.layer, glyphName, glyphName, createOptions) // glyphs will be undrawn at this point
+    for (const [layerIndex, dataPoint] of parsedData.entries()) {
+      const newGlyph = new glyphClass(
+          layerIndex,
+          namingField === '_default' ? layerIndex : dataPoint[namingField],  // use layer number as id if no naming field was specified
+          glyphName
+      )
+      if (state.project.glyphs.length === layerIndex) {
+        state.project.glyphs.push(newGlyph)
       } else {
-        emptyGlyph = new glyphClass(glyph.layer, glyphName, glyphName) // default options will be used
+        state.project.glyphs[layerIndex].registerChild(newGlyph)
       }
-      glyph.registerChild(emptyGlyph)
     }
   },
 
@@ -181,7 +175,18 @@ export default {
     glyph.changeLayer(layerId)
   },
 
-  drawGlyph: (state, {boundingRect, glyphId, dataPoint, varShapeAssignment, shapeJSONStore, glyphBoxes}) => {
+  renameGlyphs: (state, {oldName, newName}) => {
+    for (const glyph of state.project.glyphs) {
+      for (const childGlyph of glyph.iter()) {
+        if (childGlyph.name === oldName) {
+          childGlyph.name = newName
+
+        }
+      }
+    }
+  },
+
+  drawGlyph: (state, {boundingRect, glyphId, dataPoint, varShapeAssignment, shapeJSONStore}) => {
     /* params {
           glyphIndex: Number
           boundingRect: {left: Number, top: Number, width: Number, height: Number}
@@ -192,17 +197,18 @@ export default {
       if (glyph.drawn) {
         continue
       }
+      console.log(`Drawing ${glyph.name} with id ${glyphId} (layer ${glyph.layer})`)
       state.activeLayer = glyph.layer // activate layer corresponding to glyph
       glyph.activateLayer()
       // Handle scaling of glyph
       let scaleOrders = []
-      for (let binding of state.project.bindings) {
+      for (const binding of state.project.bindings) {
         if (binding.name === glyph.name) {
-          let glyphElement = state.glyphElements.find(element => element.name === binding.element)
+          const glyphElement = state.glyphElements.find(element => element.name === binding.element)
           if (glyphElement.type !== 'scale' && !glyphElement.properties.requiresTransform) {
             continue // skipping bindings that are not scales and do not require a transform
           }
-          let scaleOrder = {
+          const scaleOrder = {
             ...binding,
             value: dataPoint[binding.field]
           }
@@ -211,7 +217,7 @@ export default {
       }
       // pass custom shape according to categorical features
       const shapeJSONs = []
-      for (let assignment of varShapeAssignment) {
+      for (const assignment of varShapeAssignment) {
         if (dataPoint[assignment.field] === assignment.categoricalValue) {
           shapeJSONs.push(shapeJSONStore.get(assignment.shape).json)
           // for glyphs that require named shapes (N.B not used yet)
@@ -226,40 +232,34 @@ export default {
           }
         }
       }
-      // enforce relative position of glyph if present
-      const glyphBox = glyphBoxes.find(glyphBox => glyphBox._id === glyph.id)
       let drawOptions = {
         boundingRect: {...boundingRect},
         scaleOrders: scaleOrders,
         [state.glyphSettings.name]: state.selectedGlyphSetting,
         shapeJSONs: shapeJSONs, // used by glyphs to draw custom main paths
-        glyphBox: glyphBox
       }
       // Draw all glyph paths
       glyph.draw(drawOptions)
-      let targetGlyph
       // elements are drawn according to their priority value in decreasing order (the default is 0)
       // need to clone the binding array first as .sort() mutates its parent
       const orderedBindings = [...state.project.bindings].sort(binding => {
         let element = state.glyphElements.find(element => element.name === binding.element)
         if (element.properties.priority) {
-          return  - element.properties.priority // negative sign needed as sorting order will be ascending
+          return - element.properties.priority // negative sign needed as sorting order will be ascending
         } else {
           return 0
         }
       })
-      for (let binding of orderedBindings) {
-        let glyphElement = state.glyphElements.find(element => element.name === binding.element)
-        if (glyphElement.type === 'scale') {
-          continue // skipping bindings to scale-type elements
+      for (const binding of orderedBindings) {
+        const targetGlyph = glyph.getChild(binding.shape, 'shape')
+        const glyphElement = state.glyphElements.find(element => element.name === binding.element)
+        if (!(targetGlyph.name === binding.name && targetGlyph.shape === binding.shape)) {
+          continue // if binding is for other glyph
         }
-        if (glyph.shape === binding.shape) {
-          targetGlyph = glyph
-        } else {
-          targetGlyph = glyph.getChild(binding.shape, 'shape')
-        } // target is either main glyph or one of children
-        if (typeof targetGlyph === 'undefined') {
-          throw new Error(`Neither glyph ${glyphId} nor its children match shape '${binding.shape}' `)
+        if (glyphElement.type === 'scale') {
+          // TODO remove scale orders and let scalings be drawn here?
+          // can do, but scalings would then need to be the first bindings (can use priority element property)
+          continue // skipping bindings to scale-type elements
         }
         try {
           targetGlyph['draw' + binding.element](dataPoint[binding.field]) // get draw method for element and call it
@@ -287,17 +287,49 @@ export default {
     setTimeout(glyph.fitToBox('layer'))
   },
 
+  deleteGlyph: (state, glyphName) => { // TODO test
+    if (state.project.glyphs[0].name === glyphName) {
+      let newTopGlyph
+      for (const [idx, topGlyph] of state.project.glyphs.entries()) {
+        for (const [childIdx, childGlyph] of topGlyph.children.entries()) {
+          // find topGlyph substitute (i.e. child that is not a shape of topGlyph)
+          if (!topGlyph.constructor.shapes.children.includes(childGlyph.shape)) {
+            newTopGlyph = childGlyph
+            newTopGlyph.parent = null
+            newTopGlyph.registerItem(topGlyph.getItem('drawingBox'), 'drawingBox')
+            state.project.glyphs.splice(idx, 1, newTopGlyph)
+            topGlyph.children.splice(childIdx, 1)
+            break
+          }
+        }
+        if (newTopGlyph) {
+          topGlyph.reset({layer: false})
+          for (const childGlyph of topGlyph.children) {
+            newTopGlyph.registerChild(childGlyph)
+          }
+        } else {
+          topGlyph.reset({layer: true}) // deletes captions as well
+        }
+      }
+      state.project.glyphs = state.project.glyphs.filter(glyph => glyph.drawn)
+    } else {
+      for (const topGlyph of state.project.glyphs) {
+        topGlyph.deleteChild(glyphName)
+      }
+    }
+  },
+
   changeGlyphPosition: (state, {steps, shapeSelector = '', children = false}) => {
-    for (let glyph of state.project.glyphs) {
+    for (const glyph of state.project.glyphs) {
       let targetGlyph = glyph
       if (shapeSelector) {
          targetGlyph = [...glyph.iter()].find(glyph => glyph.name === shapeSelector)
       }
       if (children) { // must use initial glyph position
-        for (let childGlyph of targetGlyph.children) {
+        for (const childGlyph of targetGlyph.children) {
           // make child steps relative to parent glyph
-          let childSteps = steps.map(step => {
-            let childStep = JSON.parse(JSON.stringify(step)) // deep copy
+          const childSteps = steps.map(step => {
+            const childStep = JSON.parse(JSON.stringify(step)) // deep copy
             if (childStep.transform === 'shift' && childStep.parameters[2].setValues) {
               if (childStep.parameters[0] !== null) {
                 childStep.parameters[0] += childGlyph.box.shapePositions.leftShift -
@@ -315,6 +347,7 @@ export default {
       }
       targetGlyph.box.applyTransforms(steps)
       targetGlyph.fitToBox(children ? 'all' : 'glyph')
+      // else: instead of fitting to box, redrawing is trigger in the action
     }
   },
 
@@ -323,23 +356,31 @@ export default {
   }, // used in activateRedrawing action
 
   deleteElement: (state, binding) => {
-    for (let glyph of state.project.glyphs) {
-      let targetGlyph = [...glyph.iter()].find(glyph => glyph.name === binding.name)
+    for (const glyph of state.project.glyphs) {
+      const targetGlyph = [...glyph.iter()].find(glyph => glyph.name === binding.name)
       if (targetGlyph.drawn) {
+        const element = targetGlyph.constructor.elements.find(el => el.name === binding.element)
         targetGlyph.activateLayer()
-        if (targetGlyph.itemIds[binding.element.toLowerCase()]) {
+        if (element.type === 'path' && targetGlyph.itemIds[binding.element.toLowerCase()]) {
           targetGlyph.deleteItem(binding.element.toLowerCase()) // paths have same name of elements in lower-case
+        } else if (element.type === 'scale') {
+          if (element.properties.direction === 'vertical') {
+            targetGlyph.box.resize(null, 1.0, {setValues: true})
+            targetGlyph.box.toCenter()
+            targetGlyph.fitToBox()
+          } else if (element.properties.direction === 'horizontal') {
+            targetGlyph.box.resize(1.0, null, {setValues: true})
+            targetGlyph.box.toCenter()
+            targetGlyph.fitToBox()
+          }
         }
       }
     }
-    state.project.bindings = state.project.bindings.filter(
-        binding_ => !(binding_.element === binding.element && binding_.name === binding.name)
-    )
   },
 
   redrawElement: (state, {binding, normalizedData}) => {
-    for (let [i, glyph] of state.project.glyphs.entries()) {
-      let targetGlyph = [...glyph.iter()].find(glyph => glyph.name === binding.name)
+    for (const [i, glyph] of state.project.glyphs.entries()) {
+      const targetGlyph = [...glyph.iter()].find(glyph => glyph.name === binding.name)
       if (targetGlyph.drawn) {
         targetGlyph.activateLayer()
         if (targetGlyph.itemIds[binding.element.toLowerCase()]) {
@@ -362,9 +403,9 @@ export default {
       }
       glyphsToChange.push(state.project.glyphs.find(glyph => glyph.id === glyphSelector))
     }
-    for (let glyph of glyphsToChange) {
+    for (const glyph of glyphsToChange) {
       state.activeLayer = glyph.layer
-      let layer = glyphScope().project.layers[state.activeLayer]
+      const layer = glyphScope().project.layers[state.activeLayer]
       layer.activate()
       if (shapeSelector === 'layer') {
         layer.visible = value
@@ -387,7 +428,7 @@ export default {
     }
   },
 
-  resetGlyph: (state, glyphIndex) => state.project.glyphs[glyphIndex].reset(),
+  resetGlyph: (state, {glyphIndex, box}) => state.project.glyphs[glyphIndex].reset({box}),
 
   discardGlyphs: (state) => {
     state.project.glyphs = []
@@ -423,7 +464,7 @@ export default {
   // *** glyph property setter ***
   setGlyphParameters: (state, {parameters, glyphName}) => {
     let targetGlyph
-    for (let glyph of state.project.glyphs) {
+    for (const glyph of state.project.glyphs) {
       if (glyph.name === glyphName) {
         targetGlyph = glyph
       } else {
@@ -432,7 +473,7 @@ export default {
       if (!targetGlyph) {
         throw new Error(`No glyph matching '${glyphName}'`)
       }
-      for (let parameter of Object.keys(parameters)) {
+      for (const parameter of Object.keys(parameters)) {
         if (!targetGlyph.parameters.hasOwnProperty(parameter)) {
           console.warn(`Attempted to change non-existent parameter ${parameter}`)
         }
@@ -441,18 +482,40 @@ export default {
     }
   },
 
+  setGlyphColor: (state, {glyphName, fillColor, strokeColor}) => {
+    for (const glyph of state.project.glyphs) {
+      let targetGlyph
+      if (glyph.name === glyphName) {
+        targetGlyph = glyph
+      } else {
+        targetGlyph = glyph.getChild(glyphName)
+      } // target is either main glyph or one of its children
+      const mainPath = targetGlyph.mainPath
+      if (fillColor) {
+        mainPath.fillColor = fillColor
+        targetGlyph.parameters.shapeColor = fillColor
+        targetGlyph.parameters.originalFillColor = fillColor
+      }
+      if (strokeColor) {
+        mainPath.strokeColor = strokeColor
+        targetGlyph.parameters.strokeColor = strokeColor
+        targetGlyph.parameters.originalStrokeColor = strokeColor
+      }
+    }
+  },
+
   setPathParameter: (state, {parameter, value, glyphName, elementName}) => {
     let targetGlyph
-    for (let glyph of state.project.glyphs) {
+    for (const glyph of state.project.glyphs) {
       if (glyph.drawn && glyph.name === glyphName) {
         targetGlyph = glyph
       } else {
         targetGlyph = glyph.getChild(glyphName)
       } // target is either main glyph or one of its children
       if (targetGlyph && targetGlyph.drawn) {
-        let paths = targetGlyph.getNamedItems(true)
+        const paths = targetGlyph.getNamedItems(true)
         paths[elementName.toLowerCase()][parameter] = value
-        let propString = elementName.toLowerCase() + parameter[0].toUpperCase() + parameter.slice(1)
+        const propString = elementName.toLowerCase() + parameter[0].toUpperCase() + parameter.slice(1)
         targetGlyph.parameters[propString] = value // .parameters must not store objects --> assemble specific string
       }
     }
@@ -463,13 +526,13 @@ export default {
     if (glyphBoxes.length !== state.project.glyphs.length) {
       throw new Error("Passed box parameters do not match number of existing glyphs")
     }
-    for (let glyph of state.project.glyphs) {
-      let glyphBox = glyphBoxes.find(glyphBox => glyphBox._id === glyph.id)
+    for (const glyph of state.project.glyphs) {
+      const glyphBox = glyphBoxes.find(glyphBox => glyphBox._id === glyph.id)
       if (!glyphBox) {
         throw new Error(`No glyph box matching name '${glyphBox._id}'. Boxes template might not match dataset`)  // TODO apply this anyway ?
       }
       let targetGlyph
-      for (let shapeName of Object.keys(glyphBox)) {
+      for (const shapeName of Object.keys(glyphBox)) {
         if (shapeName === '_id') {
           continue // key that stores id of glyph corresponding to box
         }
@@ -503,9 +566,9 @@ export default {
   addShrinkRegrowAnimation: (state) => {
     // find out which glyphs overlap with other glyphs (must wait till this is complete to shrink glyphs)
     let glyphOverlaps = []
-    for (let glyph0 of state.project.glyphs) {
+    for (const glyph0 of state.project.glyphs) {
       let overlaps = false
-      for (let glyph1 of state.project.glyphs) {
+      for (const glyph1 of state.project.glyphs) {
         overlaps = glyph0.group.intersects(glyph1.group)
         if (overlaps) {
           console.log(`Glyph ${glyph0.id} overlaps with glyph ${glyph1.id}`)
@@ -515,7 +578,7 @@ export default {
       glyphOverlaps.push(overlaps)
     }
     // apply shrink / regrowth animation
-    for (let [i, overlap] of glyphOverlaps.entries()) {
+    for (const [i, overlap] of glyphOverlaps.entries()) {
       let glyph = state.project.glyphs[i]
       if (overlap && !glyph.animations.has('shrink-regrow')) {
         glyph.addShrinkRegrow(0.4)
@@ -524,7 +587,7 @@ export default {
   },
 
   removeShrinkRegrowAnimation: (state) => { // TODO use this when switching from chart to grid / edi
-    for (let glyph of state.project.glyphs) {
+    for (const glyph of state.project.glyphs) {
       if (glyph.animations.has('shrink-regrow')) {
         glyph.removeShrinkRegrow()
       }
